requirements.txt
streamlit
faiss-cpu
google-generativeai
beautifulsoup4
requests
langchain
unstructured
sentence-transformers
